Cleaner implementation. We are ready to become truly fast now!

+++ Meteorites.Core/Defs.h
@@ -22,7 +22,7 @@
 #include <unordered_set>
 #include <vector>
 
-typedef double real;
+typedef float real;
 
 #if !defined(M_PI)
   constexpr double M_PI = 3.1415926535897932384626433832795;
